+++ lib/busbar_cli/models/nodetype.rb
@@ -1,0 +1,12 @@
+class Nodetype
+  include Virtus.model
+
+  attribute :id, String
+  attribute :cpu, String
+  attribute :guaranteed_cpu, String
+  attribute :memory, String
+
+  def as_text
+    "#{id} =>\tCPU: #{cpu}\tMemory: #{memory}"
+  end
+end