+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## [1.9.3] - 2019-01-23
+### Add
+- Add components to busbar show
+- Add busbar nodetypes command - Fetch data of available nodetypes
+
 ## [1.9.2] - 2018-06-05
 ### Refactor
 - Refactor CLI to allow apps with names longer than 18 characters